+++ cellfinder/napari/detect/detect_containers.py
@@ -70,7 +70,7 @@ class DetectionInputs(InputContainer):
     log_sigma_size: float = 0.2
     n_sds_above_mean_thresh: float = 10
     soma_spread_factor: float = 1.4
-    max_cluster_size: int = 100000
+    max_cluster_size: float = 100000
     detection_batch_size: int = 1
 
     def as_core_arguments(self) -> dict:
